getting-started-guides/gs-rest-service/src/test/java/com/example/restservice/GreetingControllerTests.java
package com.example.restservice;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

/**
 * Test of {@link GreetingController}.
 *
 * @author guangyi
 * @since 2022-02-02
 */
@SpringBootTest
@AutoConfigureMockMvc
public class GreetingControllerTests {

    private final MockMvc mockMvc;

    @Autowired
    public GreetingControllerTests(
            MockMvc mockMvc
    ) {
        this.mockMvc = mockMvc;
    }

    @Test
    public void noParamGreetingShouldReturnDefaultMessage() throws Exception {
        this.mockMvc.perform(MockMvcRequestBuilders.get("/greeting"))
                .andDo(MockMvcResultHandlers.print())
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.jsonPath("$.content")
                        .value("Hello, World!")
                );
    }

    @Test
    public void paramGreetingShouldReturnTailoredMessage() throws Exception {
        this.mockMvc.perform(MockMvcRequestBuilders.get("/greeting")
                .param("name", "Spring Community"))
                .andDo(MockMvcResultHandlers.print())
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.jsonPath("$.content")
                        .value("Hello, Spring Community!")
                );
    }
}
